+++ src/tui.ts
@@ -3,7 +3,7 @@ import Entropy from '@entropyxyz/sdk'
 import * as config from './config'
 import { EntropyTuiOptions } from './types'
 import { logo } from './common/ascii'
-import { print } from './common/utils'
+import { jumpStartNetwork, print } from './common/utils'
 import { loadEntropy } from './common/utils-cli'
 import { EntropyLogger } from './common/logger'
 
@@ -13,6 +13,7 @@ import { entropyBalance } from './balance/interaction'
 import { entropyTransfer } from './transfer/interaction'
 import { entropyFaucet } from './faucet/interaction'
 import { entropyProgram, entropyProgramDev } from './program/interaction'
+import yoctoSpinner from 'yocto-spinner'
 
 async function setupConfig () {
   let storedConfig = await config.get()
@@ -44,25 +45,36 @@ export default function tui (entropy: Entropy, options: EntropyTuiOptions) {
     'User Programs',
   ]
 
+  const devChoices = [
+    'Jumpstart Network',
+    // 'Create and Fund Faucet(s)'
+  ]
+
+  if (options.dev) {
+    choices = [...choices, ...devChoices]
+  }
+
   // assign exit so its last
   choices = [...choices, 'Exit']
 
   main(entropy, choices, options, logger)
 }
+const loader = yoctoSpinner()
 
-async function main (entropy: Entropy, choices, options, logger: EntropyLogger) {
+async function main (entropy: Entropy, choices: string[], options: EntropyTuiOptions, logger: EntropyLogger) {
+  if (loader.isSpinning) loader.stop()
   const storedConfig = await setupConfig()
 
   // Entropy is undefined on initial install, after user creates their first account,
   // entropy should be loaded
   if (storedConfig.selectedAccount && !entropy) {
-    entropy = await loadEntropy(storedConfig.selectedAccount, options.endpoint)
+    entropy = await loadEntropy(storedConfig.selectedAccount, options.tuiEndpoint)
   }
   // If the selected account changes within the TUI we need to reset the entropy instance being used
   const currentAccount = entropy?.keyring?.accounts?.registration?.address
   if (currentAccount && currentAccount !== storedConfig.selectedAccount) {
     await entropy.close()
-    entropy = await loadEntropy(storedConfig.selectedAccount, options.endpoint);
+    entropy = await loadEntropy(storedConfig.selectedAccount, options.tuiEndpoint);
   }
   
 
@@ -88,26 +100,26 @@ async function main (entropy: Entropy, choices, options, logger: EntropyLogger) 
 
     switch (answers.choice) {
     case 'Manage Accounts': {
-      const response = await entropyAccount(options.endpoint, storedConfig)
+      const response = await entropyAccount(options.tuiEndpoint, storedConfig)
       if (response === 'exit') { returnToMain = true }
       break
     }
     case 'Register': {
-      await entropyRegister(entropy, options.endpoint, storedConfig)
+      await entropyRegister(entropy, options.tuiEndpoint, storedConfig)
       break
     }
     case 'Balance': {
-      await entropyBalance(entropy, options.endpoint, storedConfig)
+      await entropyBalance(entropy, options.tuiEndpoint, storedConfig)
         .catch(err => console.error('There was an error retrieving balance', err))
       break
     }
     case 'Transfer': {
-      await entropyTransfer(entropy, options.endpoint)
+      await entropyTransfer(entropy, options.tuiEndpoint)
         .catch(err => console.error('There was an error sending the transfer', err))
       break
     }
     case 'Sign': {
-      await entropySign(entropy, options.endpoint)
+      await entropySign(entropy, options.tuiEndpoint)
         .catch(err => console.error('There was an issue with signing', err))
       break
     }
@@ -120,17 +132,43 @@ async function main (entropy: Entropy, choices, options, logger: EntropyLogger) 
       break
     }
     case 'User Programs': {
-      await entropyProgram(entropy, options.endpoint)
+      await entropyProgram(entropy, options.tuiEndpoint)
         .catch(err => console.error('There was an error with programs', err))
       break
     }
     case 'Deploy Program': {
-      await entropyProgramDev(entropy, options.endpoint)
+      await entropyProgramDev(entropy, options.tuiEndpoint)
         .catch(err => console.error('There was an error with program dev', err))
       break
     }
+    case 'Jumpstart Network': {
+      // TO-DO: possibly move this to it's own directory similar to the other actions
+      // could create a new system directory for system/network level functionality
+      // i.e jumpstarting, deploy faucet, etc.
+      loader.text = 'Jumpstarting Network...'
+      try {
+        loader.start()
+        const jumpStartStatus = await jumpStartNetwork(entropy, options.tuiEndpoint)
+          
+        if (jumpStartStatus.isFinalized) {
+          loader.clear()
+          loader.success('Network jumpstarted!')
+          // running into an issue where the loader displays the success message but the return to main menu
+          // prompt does not display, so for now exiting process
+          process.exit(0)
+        }
+      } catch (error) {
+        loader.text = 'Jumpstart Failed'
+        loader.stop()
+        loader.clear()
+        console.error('There was an issue jumpstarting the network', error);
+        process.exit(1)
+      }
+      break
+    }
     default: {
-      throw Error(`unsupported choice: ${answers.choice}`)
+      console.error('Unsupported Action:' + answers.choice)
+      break
     }
     }
   }