+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format extends [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Version header format: `[version] Name - year-month-day (entropy-core compatibility: version [range])`
 
-## [UNRELEASED]
+## [0.0.4] Carnage - 2024-10-23 (entropy-core compatibility: 0.3.0)
 
 ### Added
 - new: 'src/flows/register/register.ts' - service file for register pure function
@@ -43,6 +43,7 @@ Version header format: `[version] Name - year-month-day (entropy-core compatibil
   - new: './src/faucet/interaction.ts' - main entrypoint for TUI flows
   - new: './src/faucet/command.ts' - main entrypoint for CLI flows
   - new: package added - yocto-spinner for adding loading spinners to the cli
+  - new: added new menu item for TUI to trigger a jumpstart to the network (needs to only be run once)
 
 ### Changed